auth_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
python-multipart==0.0.9